requirements.txt
# ==============================================================================
# Core Framework
# ==============================================================================
fastapi
uvicorn[standard]
python-multipart

# ==============================================================================
# LangChain & LangGraph
# ==============================================================================
langchain
langchain-core
langchain-community
langchain-groq
langchain-ollama
langchain-qdrant
langchain-text-splitters
langgraph
langgraph-checkpoint-postgres

# ==============================================================================
# Vector Database — Qdrant Cloud
# ==============================================================================
qdrant-client

# ==============================================================================
# Memory — PostgreSQL (LTM store + STM checkpointer)
# ==============================================================================
psycopg[binary,pool]

# ==============================================================================
# Web Search — Tavily (CRAG fallback)
# ==============================================================================
tavily-python

# ==============================================================================
# Document Processing
# ==============================================================================
pypdf
python-docx

# ==============================================================================
# Configuration & Validation
# ==============================================================================
pydantic
pydantic-settings
python-dotenv

# ==============================================================================
# HTTP Client
# ==============================================================================
httpx
